Refactored crud methods with express router
